+++ crates/apollo-compiler/src/execution/mod.rs
@@ -4,23 +4,24 @@
 //! [execution]: https://spec.graphql.org/October2021/#sec-Execution
 //! [response]: https://spec.graphql.org/October2021/#sec-Response
 
-/// A source location (line and column numbers) for a [`GraphQLError`].
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub struct GraphQLLocation {
-    /// The line number for this location, starting at 1 for the first line.
-    pub line: usize,
-    /// The column number for this location, starting at 1 and counting characters (Unicode Scalar
-    /// Values) like [`str::chars`].
-    pub column: usize,
-}
+mod response;
 
-/// A serializable error, as found in a GraphQL response.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub struct GraphQLError {
-    /// The error message.
-    pub message: String,
+pub use self::response::GraphQLError;
+pub use self::response::GraphQLLocation;
+pub use self::response::PathElement;
+pub use self::response::RequestError;
+pub use self::response::Response;
+pub use self::response::ResponseData;
+/// Re-export of the version of the `serde_json_bytes` crate used for [`JsonValue`] and [`JsonMap`]
+pub use serde_json_bytes;
 
-    /// Locations relevant to the error, if any.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub locations: Vec<GraphQLLocation>,
-}
+/// A JSON-compatible dynamically-typed value.
+///
+/// Note: [`serde_json_bytes::Value`] is similar
+/// to [`serde_json::Value`][serde_json_bytes::serde_json::Value]
+/// but uses its reference-counted [`ByteString`][serde_json_bytes::ByteString]
+/// for string values and map keys.
+pub type JsonValue = serde_json_bytes::Value;
+
+/// A JSON-compatible object/map with string keys and dynamically-typed values.
+pub type JsonMap = serde_json_bytes::Map<serde_json_bytes::ByteString, JsonValue>;